Repositories update manager
This class will be in charge of finding newest packages
from testing environment and copying these packages to
the updates environment.

This class display visual confirmation of the packages
that will be copied.

Each repository can be confiured by contraints such as
package architectures and keywords. This will be useful
to avoid copying wrong packages.

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>